// dev_gm-nvp_main-7170e7efa477/front-end/client/src/app/_models/camera.ts
export class Camera {
    camera_id: string;
    user_id: string;
    name: string;
    type:string;
    source: string;
    fps:string;
    frame: string;
    roi: Array<String>;
    gps: string;
    protocol: string;
    interface: string;
    status:string;
    agent:string;
    peopleCount: boolean;
    vehicleCount: boolean;
    vehicleTracking: boolean;
    peopleTracking: boolean;
    vehicleSpeed: boolean;
    peopleList: boolean;
    vehicleList:boolean;
    faceRecognition: boolean;
    intrusionDetection: boolean;
    faceAnalytics : boolean;
    deepAssurance : boolean;
    carCrashDetection : boolean;
    wrongTurnDetection : boolean;
    parkingViolation : boolean;
    crowdDetection : boolean;
    fashion : boolean;
    litterDetection : boolean;
    helmetDetection : boolean;
    climbingDetection : boolean;
    theme : boolean;
    anpr : boolean;
    violence : boolean;
    media_type: string;
    algo: Array<String>;
}